requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
python-multipart==0.0.12
aiofiles==24.1.0
python-dotenv==1.0.1
Pillow==10.4.0